Add solution to reverse-string
Mentor:siebenschlaefer

reverse_string() takes its argument by value but does not modify it.

F.16 of the C++ Core Guidelines says: "For “in” parameters, pass cheaply-copied types by value and others by reference to const"

In small examples like this you will not notice any difference, but if a copy is expensive or if a function is called often it can make a difference. Five years ago it was reported that half of the allocations in Google's Chrome came from std::string (link).
https://groups.google.com/a/chromium.org/forum/#!msg/chromium-dev/EUqoIz2iFU4/kPZ5ZK0K3gEJ